Create table tool bar component
